Changed order to floats

+++ alternative.go
@@ -15,7 +15,7 @@ type Alternative struct {
 	Name          string  `db:"name" json:"name" binding:"required"`
 	Description   string  `db:"description" json:"description"`
 	Cost          float32 `db:"cost" json:"cost"`
-	Order         int     `db:"order" json:"order"`
+	Order         float32     `db:"order" json:"order"`
 }
 
 // HAlternativeCreate create a ballot that belongs
